+++ app/src/androidTest/java/org/simple/clinic/overdue/AppointmentRepositoryAndroidTest.kt
@@ -2962,6 +2962,188 @@ class AppointmentRepositoryAndroidTest {
   }
 
   @Test
+  fun getting_remind_to_call_later_overdue_appointments_in_a_facility_should_work_correctly() {
+    fun createOverdueAppointment(
+        appointmentUuid: UUID,
+        patientUuid: UUID,
+        scheduledDate: LocalDate,
+        remindOn: LocalDate?,
+        facilityUuid: UUID,
+        patientAssignedFacilityUuid: UUID?,
+        patientStatus: PatientStatus,
+        appointmentStatus: Status,
+        callResult: CallResult?
+    ) {
+      val patientProfile = TestData.patientProfile(
+          patientUuid = patientUuid,
+          generatePhoneNumber = true,
+          patientAssignedFacilityId = patientAssignedFacilityUuid,
+          patientStatus = patientStatus
+      )
+      patientRepository.save(listOf(patientProfile))
+
+      val bp = TestData.bloodPressureMeasurement(
+          patientUuid = patientUuid,
+          facilityUuid = facilityUuid,
+          systolic = 120,
+          diastolic = 80
+      )
+      bpRepository.save(listOf(bp))
+
+      val bloodSugar = TestData.bloodSugarMeasurement(
+          patientUuid = patientUuid,
+          facilityUuid = facilityUuid,
+          reading = BloodSugarReading.fromMg("100", Random)
+      )
+      bloodSugarRepository.save(listOf(bloodSugar))
+
+      val appointment = TestData.appointment(
+          uuid = appointmentUuid,
+          patientUuid = patientUuid,
+          facilityUuid = facilityUuid,
+          scheduledDate = scheduledDate,
+          status = appointmentStatus,
+          cancelReason = null,
+          remindOn = remindOn
+      )
+      appointmentRepository.save(listOf(appointment))
+
+      if (callResult != null) {
+        callResultRepository.save(listOf(callResult))
+      }
+    }
+
+    //given
+    val patientWithOneDayOverdue = UUID.fromString("92eb73f2-7d15-4845-8977-6a45c62816e7")
+    val patientWithOverAnYearDaysOverdue = UUID.fromString("287558a9-3a18-49fb-8858-2bdb040412ed")
+    val deadPatientWithOverdue = UUID.fromString("fcab66bd-5e6d-4f81-9a12-c66c3db7820a")
+    val patientWithVisitedAppointment = UUID.fromString("0aa86238-51d9-4356-86cb-270b509d152d")
+    val patientWithAgreedToVisitOverdueAppointment = UUID.fromString("99628f72-29dc-4746-9f10-ec9b2cffeaed")
+    val patientWithAppointmentRemindedInFiveDaysSinceOverdue = UUID.fromString("5e84d754-3a1b-4014-856d-e6f641173407")
+    val patientWithAppointmentRemindedInTenDaysSinceOverdue = UUID.fromString("82059663-b9ff-4e56-a771-1781e70f1085")
+
+    val agreedToVisitAppointment = UUID.fromString("7c8ad2d9-2f43-4ab2-9638-263fc82d65ad")
+    val remindToCallLaterAppointmentInFiveDays = UUID.fromString("67da29f8-9d44-4996-a99d-9e8a41cdd783")
+    val remindToCallLaterAppointmentInTenDays = UUID.fromString("2a10da30-780c-455e-afaa-e8ec8a6d5cb2")
+
+    val now = LocalDate.now(clock)
+    val facility1Uuid = UUID.fromString("9759d6e0-fd4e-4b97-b43d-2099dbcb0841")
+    val facility2Uuid = UUID.fromString("0426f652-5203-484a-9ed7-c3adcbbd9f1d")
+
+    val facility1 = TestData.facility(uuid = facility1Uuid, name = "PHC Obvious")
+    val facility2 = TestData.facility(uuid = facility2Uuid, name = "PHC Bagta")
+
+    facilityRepository.save(listOf(facility1, facility2))
+
+    createOverdueAppointment(
+        appointmentUuid = UUID.fromString("cbd55cfb-9e15-4685-b698-63dae67ae55c"),
+        patientUuid = patientWithOneDayOverdue,
+        scheduledDate = now.minusDays(1),
+        remindOn = null,
+        facilityUuid = facility1Uuid,
+        patientAssignedFacilityUuid = facility1Uuid,
+        patientStatus = Active,
+        appointmentStatus = Scheduled,
+        callResult = null
+    )
+    createOverdueAppointment(
+        appointmentUuid = UUID.fromString("a2ee7bea-a519-445d-a51b-d37bd293f494"),
+        patientUuid = patientWithOverAnYearDaysOverdue,
+        scheduledDate = now.minusDays(370),
+        remindOn = null,
+        facilityUuid = facility1Uuid,
+        patientAssignedFacilityUuid = null,
+        patientStatus = Active,
+        appointmentStatus = Scheduled,
+        callResult = null
+    )
+    createOverdueAppointment(
+        appointmentUuid = UUID.fromString("0c43de19-ea9b-4f5a-ad6c-f53493b6025f"),
+        patientUuid = deadPatientWithOverdue,
+        scheduledDate = now.minusDays(7),
+        remindOn = null,
+        facilityUuid = facility1Uuid,
+        patientAssignedFacilityUuid = facility1Uuid,
+        patientStatus = Dead,
+        appointmentStatus = Scheduled,
+        callResult = null
+    )
+    createOverdueAppointment(
+        appointmentUuid = UUID.fromString("8a1463e0-5876-48ec-beaf-93723f42bbb7"),
+        patientUuid = patientWithVisitedAppointment,
+        scheduledDate = now.minusDays(30),
+        remindOn = null,
+        facilityUuid = facility1Uuid,
+        patientAssignedFacilityUuid = null,
+        patientStatus = Active,
+        appointmentStatus = Visited,
+        callResult = null
+    )
+    createOverdueAppointment(
+        appointmentUuid = agreedToVisitAppointment,
+        patientUuid = patientWithAgreedToVisitOverdueAppointment,
+        scheduledDate = now.minusDays(2),
+        remindOn = null,
+        facilityUuid = facility1Uuid,
+        patientAssignedFacilityUuid = null,
+        patientStatus = Active,
+        appointmentStatus = Scheduled,
+        callResult = TestData.callResult(
+            id = UUID.fromString("0d739533-0934-4031-9ff3-3bce3ef375df"),
+            outcome = AgreedToVisit,
+            appointmentId = agreedToVisitAppointment
+        )
+    )
+    createOverdueAppointment(
+        appointmentUuid = remindToCallLaterAppointmentInFiveDays,
+        patientUuid = patientWithAppointmentRemindedInFiveDaysSinceOverdue,
+        scheduledDate = now.minusDays(2),
+        remindOn = now.plusDays(5),
+        facilityUuid = facility1Uuid,
+        patientAssignedFacilityUuid = null,
+        patientStatus = Active,
+        appointmentStatus = Scheduled,
+        callResult = TestData.callResult(
+            id = UUID.fromString("3e84a0a7-a4ab-4c88-ba8b-5dc0072fb9f4"),
+            outcome = RemindToCallLater,
+            appointmentId = remindToCallLaterAppointmentInFiveDays
+        )
+    )
+    createOverdueAppointment(
+        appointmentUuid = remindToCallLaterAppointmentInTenDays,
+        patientUuid = patientWithAppointmentRemindedInTenDaysSinceOverdue,
+        scheduledDate = now.minusDays(2),
+        remindOn = now.plusDays(10),
+        facilityUuid = facility1Uuid,
+        patientAssignedFacilityUuid = null,
+        patientStatus = Active,
+        appointmentStatus = Scheduled,
+        callResult = TestData.callResult(
+            id = UUID.fromString("263605e9-0f38-4872-b54b-c5ff98a3e025"),
+            outcome = RemindToCallLater,
+            appointmentId = remindToCallLaterAppointmentInTenDays
+        )
+    )
+
+    //when
+    val overdueAppointments = PagingTestCase(
+        pagingSource = appointmentRepository.remindToCallLaterOverdueAppointmentsInFacility(
+            since = now,
+            facilityId = facility1Uuid
+        ),
+        loadSize = 10
+    ).loadPage()
+        .data
+        .map { it.appointment.patientUuid }
+
+    //then
+    assertThat(overdueAppointments).isEqualTo(listOf(
+        patientWithAppointmentRemindedInFiveDaysSinceOverdue,
+        patientWithAppointmentRemindedInTenDaysSinceOverdue
+    ))
+  }
+
+  @Test
   fun when_fetching_overdue_appointments_only_latest_appointments_that_are_not_visited_should_be_considered() {
     fun createOverdueAppointment(
         appointmentUuid: UUID,