+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Bump datadog gradle plugin to v1.5.1
 - Add query for getting pending appointments as a paging source
 - Add query for getting agreed to visit appointments as a paging source
+- Add query for getting remind to call later appointments as a paging source
 
 ### Fixes